# openapi-contract.yaml
openapi: 3.0.3
info:
  title: MALULEKE-KS Platform API
  version: "1.0.0"
  description: >
    Contract-first specification for the MALULEKE-KS platform.
    Derives from PLATFORM-CONSTITUTION-v1.md and BUSINESS-RULES-v1.md.
    Zod schemas are generated from this contract, never hand-written independently of it.
servers:
  - url: https://malulekeks.dev/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - {}
paths:
...
  /admin/systems/{id}:
    get:
      tags: [Systems (Admin)]
      summary: Get one system with its unmasked curation detail
      security: [{ adminSession: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Full system detail, including skills, status history, pace and activity }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Systems (Admin)]
      summary: Update a system, including publishing state transitions
      description: >
        BR-1.1 is enforced here: a request that sets contentStatus=published
        while clientVisibility != PUBLIC and clientApproved=false returns 409.
        BR-1.11: when the system's repo relationship requires the repo owner's
        permission, publishing without a granted answer returns 409
        OWNER_PERMISSION_REQUIRED. The database enforces both.
      security: [{ adminSession: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SystemUpdateInput" }
      responses:
        "200":
          description: Updated system
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SystemAdmin" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409":
          description: |
            Publishing blocked. CLIENT_APPROVAL_REQUIRED — client approval required (BR-1.1).
            OWNER_PERMISSION_REQUIRED — a collaborated system needs the repo owner's granted permission (BR-1.11).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ============================================================
  # INQUIRIES
  # ============================================================
...
  /admin/timeline:
    get:
      tags: [Timeline]
      summary: List all timeline entries for management
      security: [{ adminSession: [] }]
      responses:
        "200":
          description: Full timeline entry list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/TimelineEntry" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Timeline]
      summary: Create a timeline entry
      security: [{ adminSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TimelineCreateInput" }
      responses:
        "201":
          description: Timeline entry created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TimelineEntry" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /admin/timeline/{id}:
    patch:
      tags: [Timeline]
      summary: Update a timeline entry
      security: [{ adminSession: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TimelineCreateInput" }
      responses:
        "200":
          description: Timeline entry updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TimelineEntry" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Timeline]
      summary: Delete a timeline entry
      description: Timeline entries, unlike Systems, may be hard-deleted directly — they carry no publishing-state history worth preserving.
      security: [{ adminSession: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ============================================================
  # ADMIN — CV (Experience / Education / Skill)
  # ============================================================
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: admin_session
      description: >-
        HTTP-only, SameSite=Strict session cookie, issued only after 2FA verification (BR-3.1).
        Every state-changing admin request — and the two login steps — is refused with
        403 CSRF_REJECTED when its Origin or Sec-Fetch-Site names another site (BR-3.9).
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Missing or invalid admin session
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - NOT_FOUND
                - CLIENT_APPROVAL_REQUIRED
                - INVALID_STATUS_TRANSITION
                - RATE_LIMITED
                - ACCOUNT_LOCKED
                - CHALLENGE_EXPIRED
                - CHALLENGE_INVALID
                - LOOKUP_KEY_DEPRECATED
                - LOOKUP_KEY_EXISTS
                - OWNER_PERMISSION_REQUIRED
                - INTERNAL_ERROR
              description: Closed set, kept in lockstep with lib/schemas.ts ErrorCodeEnum — never a freeform string.
            message: { type: string }
            details: { type: object, nullable: true }
...
    OwnerPermission:
      type: string
      enum: [not_required, not_requested, requested, granted, declined]
      description: The repo owner's answer for a collaborated system (BR-1.11). A fixed control state; the database keeps it consistent with the repo relationship.
    SystemPublic:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        organization: { type: string }
        status: { type: string }
        statusColorToken: { type: string, description: "CSS custom-property/Tailwind token name (e.g. signal-finished) — status color is data (EXT-1), not a client-side hardcoded map" }
        domain: { type: string, nullable: true }
        description: { type: string }
        repoUrl: { type: string, nullable: true, description: "Null when clientVisibility = NDA_RESTRICTED (BR-1.3) or the repo is private (BR-1.7)" }
        liveUrl: { type: string, nullable: true }
        screenshotUrl: { type: string, nullable: true, description: "Homepage preview image; renders a technical placeholder when absent" }
        techStack: { type: array, items: { type: string } }
        isFlagship: { type: boolean }
        repoPrivate: { type: boolean, description: "BR-1.7 — shown as a private repository; access is available on request" }
    SystemPublicDetailed:
      allOf:
        - $ref: "#/components/schemas/SystemPublic"
        - type: object
          properties:
            caseStudyBody: { type: string }
            impacts:
              type: array
              items:
                type: object
                properties:
                  label: { type: string }
                  value: { type: string }
            testimonials:
              type: array
              items: { $ref: "#/components/schemas/TestimonialPublic" }
    TestimonialPublic:
      type: object
      description: Only ever returned where hasPermission=true (BR-6.1)
      properties:
        authorName: { type: string }
        authorRole: { type: string, nullable: true }
        organization: { type: string, nullable: true }
        quote: { type: string }
    SystemAdmin:
      allOf:
        - $ref: "#/components/schemas/SystemPublicDetailed"
        - type: object
          properties:
            contentStatus: { type: string, enum: [draft, published, archived] }
            publishAt: { type: string, format: date-time, nullable: true, description: "BR-1.13 — admin responses only. When published content goes live; null = as soon as published" }
            clientVisibility: { type: string, enum: [PUBLIC, REQUIRES_APPROVAL, NDA_RESTRICTED, ANONYMIZED_ONLY] }
            clientApproved: { type: boolean, description: "Authorizes publishing (BR-1.1) — see nameDisclosureApproved for real-name authorization" }
            nameDisclosureApproved: { type: boolean, description: "Separate from clientApproved (BR-1.4) — authorizes revealing the real org name for an ANONYMIZED_ONLY system" }
            needsCuration: { type: boolean }
            sortOrder: { type: integer }
            featuredOnHome: { type: boolean, description: "Admin-chosen for the homepage (published systems only are shown)" }
            homeOrder: { type: integer, description: "Order among homepage-featured systems, ascending" }
            onCv: { type: boolean, description: "Listed under Projects on the CV (published systems only)" }
            cvOrder: { type: integer }
            repoRelationship: { type: string, nullable: true, description: "RepoRelationship key (e.g. owner, collaborator)" }
            ownerPermission: { $ref: "#/components/schemas/OwnerPermission" }
            ownerPermissionFrom: { type: string, nullable: true, description: "Who granted or declined — the repo owner's name or GitHub login" }
            ownerPermissionAt: { type: string, format: date-time, nullable: true, description: "Set by the database when the answer changes" }
            ownerPermissionNote: { type: string, nullable: true }
            github:
              type: object
              description: Sourced from GitHub by the sync; null values until synced
              properties:
                fullName: { type: string, nullable: true }
                ownerLogin: { type: string, nullable: true }
                pushedAt: { type: string, format: date-time, nullable: true }
                languages: { type: object, nullable: true, additionalProperties: { type: integer }, description: "Bytes per language" }
                topics: { type: array, items: { type: string } }
                stars: { type: integer, nullable: true }
                syncedAt: { type: string, format: date-time, nullable: true }
...
    SystemUpdateInput:
      type: object
      properties:
        slug: { type: string, pattern: "^[a-z0-9]+(-[a-z0-9]+)*$", maxLength: 100, description: "#87 / BR-1.14 — rename the URL; the old slug keeps redirecting permanently. 409 SLUG_TAKEN (another system's current slug) or SLUG_RESERVED (another system's old slug)" }
        contentStatus: { type: string, enum: [draft, published, archived] }
        publishAt: { type: string, format: date-time, nullable: true, description: "BR-1.13 — schedule publication: with contentStatus published, the content goes live at this time (checked by every publish gate now). Omitted = unchanged; null = live as soon as published. Refused (400) on unpublished content" }
        clientApproved: { type: boolean }
        nameDisclosureApproved: { type: boolean }
        isFlagship: { type: boolean }
        sortOrder: { type: integer }
        featuredOnHome: { type: boolean }
        homeOrder: { type: integer, minimum: 0 }
        onCv: { type: boolean }
        cvOrder: { type: integer, minimum: 0 }
        repoRelationship: { type: string, nullable: true, description: "A RepoRelationship key; null clears it" }
        ownerPermission: { type: string, enum: [not_requested, requested, granted, declined] }
        ownerPermissionFrom: { type: string, nullable: true, description: "Required (by the database) when the answer is granted or declined" }
        ownerPermissionNote: { type: string, nullable: true }
        caseStudyBody: { type: string }
        repoUrl: { type: string, nullable: true }
        liveUrl: { type: string, nullable: true }
        screenshotUrl: { type: string, nullable: true }
...
    TimelineEntry:
      type: object
      properties:
        id: { type: string }
        milestoneType: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        date: { type: string, format: date }
        tags: { type: array, items: { type: string } }
        contentStatus: { type: string, enum: [draft, published, archived], description: "Public reads return published entries only" }
        publishAt: { type: string, format: date-time, nullable: true, description: "BR-1.13 — admin responses only. When published content goes live; null = as soon as published" }
        autoDrafted: { type: boolean, description: "Drafted by the database when a system first shipped; awaits the admin's approval" }
        systemId: { type: string, nullable: true }
    TimelineCreateInput:
      type: object
      required: [milestoneTypeId, title, date]
      properties:
        milestoneTypeId: { type: string }
        title: { type: string, minLength: 1 }
        description: { type: string, nullable: true }
        date: { type: string, format: date }
        media: { type: string, nullable: true }
        tags: { type: array, items: { type: string }, default: [] }
        contentStatus: { type: string, enum: [draft, published, archived], description: "Omitted on create = published; omitted on update = unchanged. Publishing an auto-drafted entry approves it." }
        publishAt: { type: string, format: date-time, nullable: true, description: "BR-1.13 — schedule publication: with contentStatus published, the content goes live at this time (checked by every publish gate now). Omitted = unchanged; null = live as soon as published. Refused (400) on unpublished content" }
...